requirements.txt
openai>=1.30.0
fastapi>=0.111.0
uvicorn>=0.29.0
pandas>=2.2.0
pyarrow>=16.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
requests>=2.32.0
overclaw
airbyte-cdk>=3.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
